removed None option in Trainer
